Develop

+++ src/users/services/users.service.ts
@@ -1,21 +1,22 @@
-import {
-  BadRequestException,
-  ConflictException,
-  Injectable,
-  NotFoundException,
-} from '@nestjs/common';
+import { BadRequestException, ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { InjectRepository } from '@nestjs/typeorm';
-import { DataSource, Repository } from 'typeorm';
+import { DataSource, ILike, Repository } from 'typeorm';
 import {
+  AccountRoleDto,
+  AccountSearchDto,
+  AccountUpdateDto,
   ChangeUserDto,
+  DeleteDto,
   RegistreUserDto,
   ResetUserDto,
+  VerifyDto,
 } from '../dtos/users.dto';
 import { Users } from '../entities/users.entity';
 
 import * as bcrypt from 'bcrypt';
 import { generateHashPassword } from 'src/common/auth/bcryptUtils';
+import { generateTokenEmail, verifyTokenEmail } from 'src/common/auth/jwtUtils';
 import { AddCronJob, EmailService } from 'src/email/services/email.service';
 import { v4 as uuidv4 } from 'uuid';
 
@@ -57,23 +58,22 @@ export class UsersService {
       const userCreate = await this.userRepo.save(newUser);
 
       this.emailServices.addCronJob({
-        type: AddCronJob.Registre,
         passwordTemporality: temporaryPassword,
+        user_id: userCreate.user_id,
+        type: AddCronJob.Registre,
         email: userCreate.email,
       });
 
       return userCreate;
     } catch (error) {
       if (error.code === '23505') {
-        throw new ConflictException(
-          `El correo electrónico ${data.email} ya se encuentra registrado`,
-        );
+        throw new ConflictException(`El correo electrónico ${data.email} ya se encuentra registrado`);
       }
       throw new NotFoundException('Error creando el usuario: ' + error.message);
     }
   }
 
-  // ! CHANGE
+  // ! CHANGE && ACCOUNT PASSWORD
   async change({ email, password, newPassword }: ChangeUserDto) {
     try {
       const user = await this.findByEmail(email);
@@ -104,9 +104,7 @@ export class UsersService {
       });
 
       if (!user) {
-        throw new BadRequestException(
-          `El correo electrónico ${email} no se encuentra registrado`,
-        );
+        throw new BadRequestException(`El correo electrónico ${email} no se encuentra registrado`);
       }
 
       const temporaryPassword: string = uuidv4().split('-', 1)[0];
@@ -117,9 +115,10 @@ export class UsersService {
       const newUser = await queryRunner.manager.save(user);
 
       await this.emailServices.addCronJob({
+        passwordTemporality: temporaryPassword,
+        user_id: newUser.user_id,
         type: AddCronJob.Reset,
         email: newUser.email,
-        passwordTemporality: temporaryPassword,
       });
 
       await queryRunner.commitTransaction();
@@ -132,8 +131,95 @@ export class UsersService {
     }
   }
 
+  // ! ACCOUNT UPDATE
+  async accountUpdate({ user_id, email, lastName, name, phone, newEmail }: AccountUpdateDto) {
+    try {
+      let verifiedEmailUpdate: boolean = true;
+
+      if (newEmail !== email) {
+        verifiedEmailUpdate = false;
+        const token = generateTokenEmail({ user_id, email: newEmail });
+        this.emailServices.addCronJob({
+          type: AddCronJob.ValidateEmail,
+          tokenJWT: token,
+          email: newEmail,
+          user_id,
+        });
+      }
+
+      const userUpdate = await this.findOne(user_id);
+      userUpdate.name = name;
+      userUpdate.lastName = lastName;
+      userUpdate.phone = phone;
+      userUpdate.verifiedEmail = verifiedEmailUpdate;
+      await this.userRepo.save(userUpdate);
+
+      return userUpdate;
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  // ! ACCOUNT SEARCH
+  async accountSearch({ search }: AccountSearchDto) {
+    try {
+      const dataUserAll = await this.userRepo.find({
+        where: [{ email: ILike(`%${search}%`) }, { name: ILike(`%${search}%`) }, { lastName: ILike(`%${search}%`) }],
+        take: 10,
+      });
+      return dataUserAll;
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  // ! UPDATE ROLES
+  async accountRoles({ user_id, role }: AccountRoleDto) {
+    try {
+      let previousRoles = '';
+      const userUpdate = await this.userRepo.findOne({ where: { user_id } });
+      if (!userUpdate) {
+        throw new BadRequestException(`Usuario no existe`);
+      }
+      previousRoles = userUpdate.roles;
+      userUpdate.roles = role;
+      await this.userRepo.save(userUpdate);
+
+      const users = await this.userRepo.find();
+
+      return { users, user: userUpdate, previousRoles };
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  // ! VERIFY
+  async verify({ token }: VerifyDto) {
+    try {
+      const decoded: { user_id: string; email: string; token?: boolean } = verifyTokenEmail(token);
+
+      const userUpdate = await this.userRepo.findOne({ where: { user_id: decoded.user_id } });
+      if (!userUpdate) {
+        throw new BadRequestException(`Token invalido, solicita nuevamente el cambio de correo electrónico`);
+      }
+      userUpdate.email = decoded.email;
+      userUpdate.verifiedEmail = true;
+      await this.userRepo.save(userUpdate);
+
+      return userUpdate;
+    } catch (error) {
+      throw error;
+    }
+  }
+
   // ! DELETE
-  remove(user_id: number) {
-    return this.userRepo.delete(user_id);
+  async remove({ user_id }: DeleteDto) {
+    const user = await this.userRepo.findOne({ where: { user_id } });
+    if (!user) {
+      throw new NotFoundException(`Usuario con ID ${user_id} no encontrado.`);
+    }
+
+    await this.userRepo.delete(user_id);
+    return user;
   }
 }
